requirements.txt
fastapi
uvicorn
pandas
faiss-cpu
sentence-transformers
cohere
python-dotenv

